move connections, regenerate openapi files
