+++ src/pages/auth/reset-password/confirm.page.tsx
@@ -8,9 +8,9 @@ import Text from "@/components/elements/Text/Text";
 import Icon, { IconNames } from "@/components/extensive/Icon/Icon";
 import BackgroundLayout from "@/components/generic/Layout/BackgroundLayout";
 import ContentLayout from "@/components/generic/Layout/ContentLayout";
+import { useRequestPassword } from "@/connections/ResetPassword";
 import { requestPasswordReset } from "@/generated/v3/userService/userServiceComponents";
 import { useValueChanged } from "@/hooks/useValueChanged";
-import { useRequestPassword } from "@/store/apiSlice";
 
 const SignupConfirmPage = ({ email }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
   const t = useT();
